Add files via upload
With the continuous expansion of urban areas, water consumption and drainage in cities have been on the rise, leading to water resource scarcity and water quality degradation. The growing prominence of water-related issues has significantly hindered China's social and economic development. Among these challenges, domestic sewage treatment in urban areas stands out as a pressing concern, requiring valuable community space and substantial financial investment for cesspool construction and maintenance.

To alleviate the financial burden on communities, we've developed a mathematical model for a non-homogeneous flow cesspool. In contrast to the uniform flow cesspool model mentioned in the title, our first innovation introduces non-uniform output, where the cesspool's discharge varies with the incoming water per unit of time.

Without considering sedimentation in the cesspool, we initially devised Model I. Over a 30-day period (720 hours), we generated 720 random numbers to simulate the hourly water inflow into the cesspool. Let f(t) represent the inflow at time t, and g(m) denote the discharge in the mth hour, ultimately yielding the capacity of the cesspool at time t, denoted as c(t). The calculated flow volume for inflow and outflow can be derived from the time-based graph. Our results demonstrate that, compared to the original uniform flow cesspool, this model only occupies 31.1% of the original space, greatly improving space utilization and illustrating the practical significance of our model.

To make our model more realistic, our second innovation extends Model I by adding sedimentation in the water, resulting in Model II. In this case, we developed an optimization model to define an objective function for maximizing capacity while considering an objective cost function. The objective cost function encompasses construction, maintenance, and cleaning costs. We assume that sediment cannot be discharged with wastewater at any time, as doing so may lead to pipe clogging and reduced drainage efficiency. We consider that sewage layers after entering the cesspool, each layer's retention time is observed, and when the retention time reaches 24 hours, we conclude that the water layer has settled. After solving the objective cost function, we find that the total cost is minimized when water is deposited for 23 hours per day, maximizing capacity at that time.

Lastly, we conducted a sensitivity analysis on the influence of different sand content levels on the optimal deposition time, concluding that sand content between 0.4 and 0.5 has the most significant impact on the results.